refactor the code to make class clean
