+++ Sources/BaseComponents/Network/NetFetch.swift
@@ -16,93 +16,7 @@
 
 import Foundation
 
-public class NetFetchResponse {
-    public var data: Data?
-    public var urlRequest: URLRequest?
-    public var urlResponse: URLResponse?
-    public var error: Error?
-    public var url: URL?
-    public var urlString: String?
-    
-    public func string(encoding: String.Encoding = .utf8) -> String {
-        if let data = data {
-            return String(data: data, encoding: .utf8) ?? ""
-        }
-        return ""
-    }
-    
-    public func bind<T: Codable>(_ modelStruct: T.Type) -> T? {
-        guard let data = data else {
-            return nil
-        }
-        
-        do {
-            return try JSONDecoder().decode(modelStruct, from: data)
-        } catch let parsingError {
-            print("Binding error: ", parsingError)
-        }
-        
-        return nil;
-    }
-}
 
-public class NetFetchRequest {
-    public let urlString: String
-    public let completionHandler: ((NetFetchResponse)->())
-    
-    public var httpMethod: String = "GET"
-    public var parameters: Dictionary<String, String>? = nil
-    public var body: Data? = nil
-    public var headers: Dictionary<String, String>? = nil
-    public var timeoutInterval: TimeInterval = 30.0
-    public var cachePolicy: URLRequest.CachePolicy = .useProtocolCachePolicy
-    public var retryOnFailure = false
-    public var ignoreQueue = false
-    public weak var dataTask: URLSessionDataTask?
-    
-    public init(urlString: String, completionHandler: @escaping((NetFetchResponse)->())) {
-        self.urlString = urlString
-        self.completionHandler = completionHandler
-    }
-    
-    public func urlRequest() -> URLRequest? {
-        var urlComponents = URLComponents(string: urlString)
-        if let parameters = parameters {
-            var queryItems: Array<URLQueryItem> = Array()
-            parameters.forEach {
-                queryItems.append(URLQueryItem(name: $0, value: $1))
-            }
-            urlComponents?.queryItems = queryItems
-        }
-        guard let url = urlComponents?.url else {
-            return nil
-        }
-        var urlRequest = URLRequest(url: url, cachePolicy: cachePolicy, timeoutInterval: timeoutInterval)
-        if let headers = headers {
-            headers.forEach {
-                urlRequest.setValue($1, forHTTPHeaderField: $0)
-            }
-        }
-        urlRequest.httpMethod = httpMethod
-        if let body = body {
-            urlRequest.httpBody = body
-        }
-        return urlRequest
-    }
-    
-    public func cancel() {
-        if let task = dataTask {
-            if task.state == .running {
-                task.cancel()
-            }
-            if ignoreQueue {
-                return
-            }
-        }
-        
-        NetFetch.removeRequest(self)
-    }
-}
 
 public class NetFetch {
     static private let session = URLSession(configuration: .default)
