+++ Sources/BaseComponents/Network/NetFetchResponse.swift
@@ -1,0 +1,48 @@
+
+//
+//  NetFetchResponse.swift
+//  BaseComponents
+//
+//  Created by Marc Steven on 4.1.2020.
+//  Copyright © 2019 Maximilian Mackh. All rights reserved.
+
+/*
+ Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+
+ The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+
+ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+
+import Foundation
+
+public class NetFetchResponse {
+    public var data: Data?
+    public var urlRequest: URLRequest?
+    public var urlResponse: URLResponse?
+    public var error: Error?
+    public var url: URL?
+    public var urlString: String?
+    
+    public func string(encoding: String.Encoding = .utf8) -> String {
+        if let data = data {
+            return String(data: data, encoding: .utf8) ?? ""
+        }
+        return ""
+    }
+    
+    public func bind<T: Codable>(_ modelStruct: T.Type) -> T? {
+        guard let data = data else {
+            return nil
+        }
+        
+        do {
+            return try JSONDecoder().decode(modelStruct, from: data)
+        } catch let parsingError {
+            print("Binding error: ", parsingError)
+        }
+        
+        return nil;
+    }
+}
